mailin
Module response

Response contains a selection of SMTP responses for use in handlers.

Structs

Response

Response contains a code and message to be sent back to the client

Enums

Action

Action indicates the recommended action to take on a response

Constants

AUTHENTICATION_REQUIRED

Authentication required

AUTH_OK

Authentication succeeded

BAD_HELLO

Error with HELO

BAD_MAILBOX

Invalid mailbox name

BLOCKED_IP

IP address on blocklists

GOODBYE

Response to indicate that the SMTP session finished

INTERNAL_ERROR

Internal server error

INVALID_CREDENTIALS

Bad authentication attempt

NO_MAILBOX

Unknown user

NO_SERVICE

Service not available

NO_STORAGE

User storage quota exceeded

OK

OK response

OUT_OF_SPACE

Insufficient system storage

START_DATA

Response sent to the client before accepting data

TEMP_AUTH_FAILURE

Authentication system is not working

TRANSACTION_FAILED

Error handling incoming message